+++ mod-batteries.ts
@@ -2,5 +2,5 @@
 
 export { HtmlShell } from "./src/HtmlShell.ts";
 export { byoHtml } from "./src/ssr.ts";
-export { byoHat, Hat } from "./src/miscellaneous/Hat.ts";
+export { Hat, makeHat } from "./src/miscellaneous/Hat.ts";
 export * from "./src/minifyCss.ts";